Add Exporter Class
Exporter is a class whose purpoes is to export the loose-association produced by Loose.associate. It's a collaborator of Loose but it has been externalized to a class in order to separate the tasks. Loose knows the logic, while Exporter knows nothing about the logic but it just export to different formats.

Exporter uses DB-API for relational databases. In this way it's easy to export to a wide variety of relational DB without changing the to_db method.

The method to_sqlite connects to a sqlite database and the calls to_db to export.

Exporter automatically creates a table for every fragment using the correct names and types. It also creates an index for the group attribute.

A unique index on all the groups is used for the associations table (to check once more the association heterogenity property).

Exporter also export a schema table which is a catalog of which attributes are stored in every fragment (used in next commits).